Modules
Moved code from main to modules.
I think using super:: all the time is a bad idea,
but I'll have to learn how the module system works first.
